Create new util for retrieving device id
