+++ src/js/utils/device.js
@@ -1,0 +1,11 @@
+import { storage } from 'storage';
+
+export function getDeviceId() {
+    const SK_STORAGE = 'sk_deviceid';
+    const deviceId = storage.getItem(SK_STORAGE) ||
+    uuid.v4().replace(/-/g, '');
+
+    storage.setItem(SK_STORAGE, deviceId);
+
+    return deviceId;
+}
